feat: Add AI chat component with modal functionality and update dependencies
