+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/shuding/nextra-docs-template#readme",
   "dependencies": {
+    "@google/generative-ai": "^0.21.0",
     "next": "^14.2.17",
     "nextra": "2.13.4",
     "nextra-theme-docs": "2.13.4",
